+++ src/fr/dwaps/model/dao/InterfaceDAO.java
@@ -4,8 +4,8 @@ import java.util.List;
 
 public interface InterfaceDAO<T> {
 	List<T> findAll();
-	T find(long id);
-	void delete(long id);
+	T find(int id);
+	void delete(int id);
 	void update(T obj);
 	void create(T obj);
 }
